split node search
