+++ src/node-search/context.js
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const Context = React.createContext({});
