--- src/reducers.js
@@ -1,91 +1,0 @@
-import { transferObjectProps } from './util/object.js';
-import { transferQueryProps } from './util/object.js';
-import { copyObject } from './util/object.js';
-
-// map previous global state to new global state based on action
-export function Reducer(prevState, action) {
-  // start with previous state
-  let newState;
-  if (prevState === undefined)
-    newState = {};
-  else
-    newState = { ...prevState };
-
-  // detect action type and set new state accordingly
-  switch (action.type) {
-    // set definitions
-    case 'set_definitions':
-      newState.metagraph = copyObject(action.payload.metagraph) || {};
-      newState.hetioDefinitions =
-        copyObject(action.payload.hetioDefinitions) || {};
-      newState.hetioStyles = copyObject(action.payload.hetioStyles) || {};
-      newState.hetmechDefinitions =
-        copyObject(action.payload.hetmechDefinitions) || {};
-      break;
-
-    // update source and/or target node
-    case 'update_source_target_nodes':
-      if (action.payload.sourceNode !== undefined)
-        newState.sourceNode = copyObject(action.payload.sourceNode) || {};
-      if (action.payload.targetNode !== undefined)
-        newState.targetNode = copyObject(action.payload.targetNode) || {};
-      break;
-
-    // swap source/target nodes
-    case 'swap_source_target_nodes':
-      newState.sourceNode = copyObject(prevState.targetNode) || {};
-      newState.targetNode = copyObject(prevState.sourceNode) || {};
-      break;
-
-    // update metapaths
-    case 'update_metapaths':
-      newState.metapaths = copyObject(action.payload.metapaths) || [];
-      if (action.preserveChecks === true) {
-        transferObjectProps(
-          prevState.metapaths,
-          newState.metapaths,
-          ['id'],
-          ['checked']
-        );
-      }
-      break;
-
-    // update path queries
-    case 'update_path_queries':
-      newState.pathQueries = copyObject(action.payload.pathQueries) || [];
-      if (action.preserveChecks === true) {
-        transferQueryProps(
-          prevState.pathQueries,
-          newState.pathQueries,
-          'paths',
-          ['node_ids', 'rel_ids'],
-          ['checked', 'highlighted']
-        );
-      }
-      break;
-
-    default:
-      break;
-  }
-
-  // make sure critical variables defined
-  if (!newState.metagraph)
-    newState.metagraph = {};
-  if (!newState.hetioDefinitions)
-    newState.hetioDefinitions = {};
-  if (!newState.hetioStyles)
-    newState.hetioStyles = {};
-  if (!newState.hetmechDefinitions)
-    newState.hetmechDefinitions = {};
-  if (!newState.sourceNode)
-    newState.sourceNode = {};
-  if (!newState.targetNode)
-    newState.targetNode = {};
-  if (!newState.metapaths)
-    newState.metapaths = [];
-  if (!newState.pathQueries)
-    newState.pathQueries = [];
-
-  // set new state
-  return newState;
-}
